Move Card list location
